backend/requirements.txt
starlette==0.41.3
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.0.1
